fix: split CLI option descriptions
