+++ packages/cli/src/cli/command-option-descriptions.ts
@@ -1,0 +1,110 @@
+const descriptions: Record<string, string> = {
+  "--add-npm-scripts": "Add npm script shortcuts during initialization.",
+  "--allow-dirty": "Allow running while the working tree has changes.",
+  "--allow-open-findings": "Allow superseding work with unresolved findings.",
+  "--allow-scripts": "Allow preset script execution.",
+  "--apply": "Apply the operation instead of planning it.",
+  "--archive": "Archive generated governance output.",
+  "--archive-field": "Set the field used for archive disposition.",
+  "--archived-by": "Record the actor archiving the task.",
+  "--arbiter": "Set the decision arbiter as agent:<id>, human:<id>, or system:<id>.",
+  "--assume-locale": "Set the assumed locale for migrated content.",
+  "--base": "Set the git base ref.",
+  "--branch": "Set the module branch.",
+  "--body": "Set authored body content for the generated decision document.",
+  "--by": "Supersede by an existing task id.",
+  "--catalog": "Use a template catalog file.",
+  "--chosen": "Set the selected decision option text.",
+  "--ci": "Set the completion CI gate result.",
+  "--claim": "Set the primary supporting claim text for a decision.",
+  "--confirm": "Confirm a destructive or relation-changing action.",
+  "--confirm-plan": "Confirm a migration plan before applying it.",
+  "--current-step": "Set the current module step.",
+  "--decided-at": "Set the decision timestamp for transition commands.",
+  "--deleted-by": "Record the actor deleting or superseding the task.",
+  "--depends-on": "Register a module dependency.",
+  "--dry-run": "Preview the operation without writing changes.",
+  "--evidence": "Attach evidence in type:path:summary format.",
+  "--evidence-relation": "Attach a decision anchor to a task or fact evidence ref as anchor:type:target:rationale; repeat for multiple relations.",
+  "--confidence": "Set fact confidence as low, medium, or high.",
+  "--approval": "Record a runtime event approval decision.",
+  "--at": "Set the runtime event timestamp.",
+  "--interrupt": "Record a task-level interrupt or steering action.",
+  "--kind": "Set the runtime event kind.",
+  "--result": "Record a runtime event result status.",
+  "--runtime": "Set the observed runtime kind.",
+  "--session": "Set the runtime session id.",
+  "--step": "Set the observed runtime step id.",
+  "--summary": "Set a short structured summary.",
+  "--tool": "Set the observed tool or command name.",
+  "--total-tokens": "Set the observed total token count.",
+  "--turn": "Set the observed runtime turn id.",
+  "--force": "Force the lifecycle transition with audit metadata.",
+  "--from-legacy": "Create from a legacy task id.",
+  "--hard": "Hard-delete the selected task.",
+  "--help": "Show help output.",
+  "--id": "Set the explicit entity id when the command supports one.",
+  "--include-archived": "Include archived task packages.",
+  "--input": "Provide one script input as key=value; repeat for multiple inputs.",
+  "--json": "Emit CommandReceipt/v1 JSON.",
+  "--kernel-version": "Validate against a kernel version.",
+  "--lesson": "Filter by lesson state.",
+  "--limit": "Limit the number of planned items.",
+  "--locale": "Set generated content locale.",
+  "--long-running": "Mark the task as long-running.",
+  "--memory-class": "Classify fact memory as semantic, episodic, or procedural.",
+  "--memory-tag": "Attach a fact memory tag; repeat or comma-separate values.",
+  "--missing-materials": "Filter tasks missing required materials.",
+  "--module": "Select a registered module key; use module list to discover keys.",
+  "--module-scope": "Set the registered module source scope, such as packages/name/**.",
+  "--module-title": "Set the human-readable title for a registered module.",
+  "--name": "Set the project name written to harness.yaml.",
+  "--observed-at": "Set the observation timestamp for a fact record.",
+  "--out": "Write the generated plan to a file.",
+  "--out-dir": "Set the output directory.",
+  "--owner": "Set the module owner.",
+  "--plan": "Plan without applying changes.",
+  "--plan-only": "Create a migration plan without applying it.",
+  "--post-merge": "Run checks intended for post-merge validation.",
+  "--prefix": "Set the module id prefix.",
+  "--preset": "Select a preset id; new-task defaults to standard-task and preset list shows installed presets.",
+  "--product-line": "Attach a comma-separated product line list to a decision.",
+  "--profile": "Select a check or task profile; new-task defaults to baseline.",
+  "--project": "Use the project preset layer.",
+  "--proposed-by": "Set the decision proposer as agent:<id>, human:<id>, or system:<id>.",
+  "--purpose": "Filter script entries by declared purpose.",
+  "--question": "Set the decision question being answered.",
+  "--queue": "Filter by queue.",
+  "--reason": "Record the reason for the lifecycle change.",
+  "--rejected": "Set a rejected decision option text.",
+  "--register-module": "Register a module while creating the task.",
+  "--review": "Filter by review state.",
+  "--reviewer": "Set the reviewer id.",
+  "--risk-tier": "Set decision risk tier as low, medium, or high.",
+  "--search": "Search task metadata and prose.",
+  "--session-dir": "Set the migration session directory.",
+  "--shared": "Register a shared path for the module.",
+  "--slug": "Set the task slug.",
+  "--scope": "Set the module scope.",
+  "--soft": "Soft-delete the selected task.",
+  "--source": "Set the fact evidence source or filter extension entries by source layer.",
+  "--state": "Set or filter by state.",
+  "--statement": "Set the fact statement text.",
+  "--status": "Set the external or module status.",
+  "--strict": "Run strict checks.",
+  "--task": "Set the task id.",
+  "--text": "Progress text appended as-is (no Markdown formatting or normalization).",
+  "--title": "Set the required task title used for generated package metadata and slug.",
+  "--urgency": "Set decision urgency as low, medium, or high.",
+  "--vertical": "Select a vertical definition; new-task defaults to software/coding.",
+  "--version": "Print the installed CLI version.",
+  "--why-not": "Set the rationale for rejecting the alternative."
+};
+
+export function optionDescription(flag: string): string {
+  const description = descriptions[flag];
+  if (!description) {
+    throw new Error(`missing CLI help option description: ${flag}`);
+  }
+  return description;
+}
