+++ tools/check-cli-help-contract.test.mjs
@@ -90,6 +90,7 @@ function writeRegistry(rootDir, body, receiptBody = "") {
   const dir = path.join(rootDir, "packages/cli/src/cli");
   mkdirSync(dir, { recursive: true });
   writeFileSync(path.join(dir, "command-registry.ts"), body);
+  writeFileSync(path.join(dir, "command-option-descriptions.ts"), body);
   writeFileSync(path.join(dir, "receipt.ts"), receiptBody);
   writeFileSync(path.join(dir, "receipt-contracts.ts"), body);
   writeFileSync(path.join(rootDir, "packages/cli/src/index.ts"), "console.log(JSON.stringify(output));\n");
